+++ History.md
@@ -22,6 +22,7 @@
 - Amendments can now be further restricted to only affect one particular location within one paragraph.
 - There are new motion statuses "Quorum reached" and "Quorum missed".
 - The CAPTCHA system is reset for an user after a successful login. This solves issues when multiple users are behind the same IP address.
+- The full-screen projector now also shows the initiator and status of a motion / amendment.
 - A new e-mail-sending library (Symfony mailer) is used. Amazon SES can now be configured as mailer, too (by editing the config.json directly).
 
 ### Version 4.9.1 (2022-02-12)